rename paths in tests
