+++ lighthouse-core/scripts/build-report-for-autodeployment.js
@@ -12,7 +12,7 @@ const fs = require('fs');
 const path = require('path');
 const swapLocale = require('../lib/i18n/swap-locale.js');
 
-const ReportGenerator = require('../../lighthouse-core/report/report-generator.js');
+const ReportGenerator = require('../../lighthouse-report/report-generator.js');
 const {defaultSettings} = require('../config/constants.js');
 const lighthouse = require('../index.js');
 const lhr = /** @type {LH.Result} */ (require('../../lighthouse-core/test/results/sample_v2.json'));
