+++ lighthouse-core/scripts/i18n/collect-strings.js
@@ -14,7 +14,7 @@ const path = require('path');
 const assert = require('assert').strict;
 const tsc = require('typescript');
 const MessageParser = require('intl-messageformat-parser').default;
-const Util = require('../../report/html/renderer/util.js');
+const Util = require('../../../lighthouse-report/renderer/util.js');
 const {collectAndBakeCtcStrings} = require('./bake-ctc-to-lhl.js');
 const {pruneObsoleteLhlMessages} = require('./prune-obsolete-lhl-messages.js');
 const {countTranslatedMessages} = require('./count-translated.js');
