+++ lighthouse-report/test/renderer/dom-test.js
@@ -8,12 +8,12 @@
 const assert = require('assert').strict;
 const fs = require('fs');
 const jsdom = require('jsdom');
-const DOM = require('../../../../report/html/renderer/dom.js');
-const Util = require('../../../../report/html/renderer/util.js');
-const I18n = require('../../../../report/html/renderer/i18n.js');
+const DOM = require('../../renderer/dom.js');
+const Util = require('../../renderer/util.js');
+const I18n = require('../../renderer/i18n.js');
 
 const TEMPLATE_FILE = fs.readFileSync(__dirname +
-    '/../../../../report/html/templates.html', 'utf8');
+    '/../../templates.html', 'utf8');
 
 /* eslint-env jest */
