+++ lighthouse-report/test/renderer/i18n-test.js
@@ -6,13 +6,13 @@
 'use strict';
 
 const assert = require('assert').strict;
-const Util = require('../../../../report/html/renderer/util.js');
-const I18n = require('../../../../report/html/renderer/i18n.js');
+const Util = require('../../renderer/util.js');
+const I18n = require('../../renderer/i18n.js');
 
 // Require i18n to make sure Intl is polyfilled in Node without full-icu for testing.
 // When Util is run in a browser, Intl will be supplied natively (IE11+).
 // eslint-disable-next-line no-unused-vars
-const i18n = require('../../../../lib/i18n/i18n.js');
+const i18n = require('../../../lighthouse-core/lib/i18n/i18n.js');
 
 const NBSP = '\xa0';
