+++ lighthouse-treemap/types/treemap.d.ts
@@ -1,9 +1,9 @@
 import _TreemapUtil = require('../app/src/util.js');
-import _TextEncoding = require('../../lighthouse-core/report/html/renderer/text-encoding.js');
+import _TextEncoding = require('../../lighthouse-report/renderer/text-encoding.js');
 import _DragAndDrop = require('../../lighthouse-viewer/app/src/drag-and-drop.js');
 import _FirebaseAuth = require('../../lighthouse-viewer/app/src/firebase-auth.js');
 import _GithubApi = require('../../lighthouse-viewer/app/src/github-api.js');
-import _Logger = require('../../lighthouse-core/report/html/renderer/logger.js');
+import _Logger = require('../../lighthouse-report/renderer/logger.js');
 import {FirebaseNamespace} from '@firebase/app-types';
 
 declare global {
