+++ types/i18n.d.ts
@@ -4,7 +4,7 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-import Util = require('../lighthouse-core/report/html/renderer/util.js');
+import Util = require('../lighthouse-report/renderer/util.js');
 
 declare global {
   module LH {
